+++ Adollib/Scripts/fall_object_manager.cpp
@@ -78,13 +78,13 @@ void Fall_object_manager::update()
 	while (timer > fall_cool_time) {
 
 		int type_num = rand() % 3;
-		Vector3 rotate_num = Vector3(rand() % 360, rand() % 360, rand() % 360);
-		int scale_num = (rand() % 100) * 0.01f * 3 + 2.2;
-		Vector3 pos = Vector3(start_pos_x[rand() % 4] + rand() % 100 * 0.01f * start_pos_x_randam_size, start_pos_y + rand() % 100 * 0.01f * start_pos_x_randam_size, -(1 + scale_num));
+		Vector3 rotate_num = Vector3((float)(rand() % 360), (float)(rand() % 360), (float)(rand() % 360));
+		int scale_num = (float)(rand() % 100) * 0.01f * 3 + 2.2;
+		Vector3 pos = Vector3(start_pos_x[rand() % 4] + (float)(rand() % 100) * 0.01f * start_pos_x_randam_size, start_pos_y + (float)(rand() % 100) * 0.01f * start_pos_x_randam_size, -(1 + scale_num));
 
-		if (type_num == 0)set_box("fall_box",    pos, rotate_num, Vector3(scale_num), Al_Global::get_gaming(rand(), 2, 196 / 255.0f, 1), fall_object_pearent);
-		if (type_num == 1)set_sphere("fall_box", pos, scale_num, Al_Global::get_gaming(rand(), 2, 196 / 255.0f, 1), fall_object_pearent);
-		if (type_num == 2)set_sphere("fall_box", pos, scale_num, Al_Global::get_gaming(rand(), 2, 196 / 255.0f, 1), fall_object_pearent);
+		if (type_num == 0)set_box("fall_box",    pos, rotate_num, Vector3(scale_num), Al_Global::get_gaming((float)rand(), 2, 196 / 255.0f, 1), fall_object_pearent);
+		if (type_num == 1)set_sphere("fall_box", pos, scale_num, Al_Global::get_gaming((float)rand(), 2, 196 / 255.0f, 1), fall_object_pearent);
+		if (type_num == 2)set_sphere("fall_box", pos, scale_num, Al_Global::get_gaming((float)rand(), 2, 196 / 255.0f, 1), fall_object_pearent);
 
 		timer = 0;
 	}
